backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
groq
datasets
pandas
python-dotenv
pydantic
huggingface_hub
